feat: merge `utils` to `core`
