+++ src/core/index.ts
@@ -2,6 +2,9 @@ export * from './processor/index.js';
 export * from './standard/index.js';
 
 export * as compare from './compare/index.js';
+export * as random from './random/index.js';
+export * as dt from './temporal/index.js';
+
 export { default as curry } from './lambda/curry.js';
 export { default as pipe } from './lambda/pipe.js';
 export { default as unique } from './standard/unique.js';
