+++ src/core/random/index.ts
